fixed sonar changes
