--- backend/homework/homework-3/src/main/java/kdu/Ques2/HealthInsurancePlan.java
@@ -1,31 +1,0 @@
-package kdu.Ques2;
-
-public class HealthInsurancePlan {
-//    private InsuranceBrand offeredBy;
-
-    private double coverage;
-    private double discount;
-
-    public double getDiscount() {
-        return discount;
-    }
-
-    public void setDiscount(double discount) {
-        this.discount = discount;
-    }
-
-    public double getCoverage() {
-        return coverage;
-    }
-
-    public void setCoverage(double coverage) {
-        this.coverage = coverage;
-    }
-
-//    public InsuranceBrand getOfferedBy() {
-//        return offeredBy;
-//    }
-//    public void setOfferedBy(InsuranceBrand offeredBy) {
-//        this.offeredBy = offeredBy;
-//    }
-}
