--- backend/homework/homework-3/src/main/java/kdu/Ques2/InsurencePlan/BronzePlan.java
@@ -1,10 +1,0 @@
-package kdu.Ques2.InsurencePlan;
-
-import kdu.Ques2.HealthInsurancePlan;
-
-public class BronzePlan extends HealthInsurancePlan {
-    public BronzePlan(){
-        setCoverage(0.6);
-        setDiscount(25.0);
-    }
-}
