--- backend/homework/homework-3/src/main/java/kdu/Ques2/InsurencePlan/GoldPlan.java
@@ -1,10 +1,0 @@
-package kdu.Ques2.InsurencePlan;
-
-import kdu.Ques2.HealthInsurancePlan;
-
-public class GoldPlan extends HealthInsurancePlan {
-    public GoldPlan(){
-        setCoverage(0.8);
-        setDiscount(40.0);
-    }
-}
