--- backend/homework/homework-3/src/main/java/kdu/Ques2/InsurencePlan/PlatinumPlan.java
@@ -1,10 +1,0 @@
-package kdu.Ques2.InsurencePlan;
-
-import kdu.Ques2.HealthInsurancePlan;
-
-public class PlatinumPlan extends HealthInsurancePlan {
-    public PlatinumPlan(){
-        setCoverage(0.9);
-        setDiscount(50.0);
-    }
-}
