--- backend/homework/homework-3/src/main/java/kdu/Ques2/InsurencePlan/SilverPlan.java
@@ -1,10 +1,0 @@
-package kdu.Ques2.InsurencePlan;
-
-import kdu.Ques2.HealthInsurancePlan;
-
-public class SilverPlan extends HealthInsurancePlan {
-    public SilverPlan(){
-        setCoverage(0.7);
-        setDiscount(30.0);
-    }
-}
