--- backend/homework/homework-3/src/main/java/kdu/Ques3/InsurencePlan/BronzePlan.java
@@ -1,15 +1,0 @@
-package kdu.Ques3.InsurencePlan;
-
-import kdu.Ques3.HealthInsurancePlan;
-
-public class BronzePlan extends HealthInsurancePlan {
-    public BronzePlan(){
-        setCoverage(0.6);
-        setDiscount(25.0);
-    }
-
-    @Override
-    public double computeMonthlyPremium(double salary) {
-        return 0.05 * salary;
-    }
-}
