--- backend/homework/homework-3/src/main/java/kdu/Ques3/InsurencePlan/SilverPlan.java
@@ -1,15 +1,0 @@
-package kdu.Ques3.InsurencePlan;
-
-import kdu.Ques3.HealthInsurancePlan;
-
-public class SilverPlan extends HealthInsurancePlan {
-    public SilverPlan(){
-        setCoverage(0.7);
-        setDiscount(30.0);
-    }
-
-    @Override
-    public double computeMonthlyPremium(double salary) {
-        return 0.06 * salary;
-    }
-}
