--- backend/homework/homework-3/src/main/java/kdu/Ques4/InsuranceBrand.java
@@ -1,5 +1,0 @@
-package kdu.Ques4;
-
-public interface InsuranceBrand {
-    public double computeMonthlyPremium(HealthInsurancePlan insurancePlan, int age, boolean smoking);
-}
