--- backend/homework/homework-3/src/main/java/kdu/Ques4/User.java
@@ -1,92 +1,0 @@
-package kdu.Ques4;
-
-import kdu.Ques4.HealthInsurancePlan;
-
-public class User {
-    private long id;
-    // add rest of the variables
-    private String firstName;
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public boolean isSmoking() {
-        return smoking;
-    }
-
-    public void setSmoking(boolean smoking) {
-        this.smoking = smoking;
-    }
-
-    private int age;
-    private boolean smoking;
-    private String lastName;
-    private String email;
-    private String gender;
-
-    private boolean insured;
-
-    private HealthInsurancePlan insurancePlan;
-
-    public HealthInsurancePlan getInsurancePlan() {
-        return insurancePlan;
-    }
-
-    public void setInsurancePlan(HealthInsurancePlan insurancePlan) {
-        this.insurancePlan = insurancePlan;
-    }
-
-    public boolean isInsured() {
-        return insured;
-    }
-
-    public void setInsured(boolean insured) {
-        this.insured = insured;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-}
-
