+++ backend/homework/homework-3/src/main/java/kdu/ques1/Doctor.java
@@ -1,0 +1,22 @@
+package kdu.ques1;
+
+public class Doctor extends Staff {
+    private long doctorId;
+    private String specialization;
+
+    public long getDoctorId() {
+        return doctorId;
+    }
+
+    public void setDoctorId(long doctorId) {
+        this.doctorId = doctorId;
+    }
+
+    public String getSpecialization() {
+        return specialization;
+    }
+
+    public void setSpecialization(String specialization) {
+        this.specialization = specialization;
+    }
+}
