+++ backend/homework/homework-3/src/main/java/kdu/ques2/InsurencePlan/BronzePlan.java
@@ -1,0 +1,10 @@
+package kdu.ques2.InsurencePlan;
+
+import kdu.ques2.HealthInsurancePlan;
+
+public class BronzePlan extends HealthInsurancePlan {
+    public BronzePlan(){
+        setCoverage(0.6);
+        setDiscount(25.0);
+    }
+}
