+++ backend/homework/homework-3/src/main/java/kdu/ques3/Patient.java
@@ -1,0 +1,16 @@
+package kdu.ques3;
+
+public class Patient extends User {
+    private long patientId;
+
+    public long getPatientId() {
+        return patientId;
+    }
+
+    public void setPatientId(long patientId) {
+        this.patientId = patientId;
+    }
+
+
+
+}
