+++ backend/homework/homework-3/src/main/java/kdu/ques4/InsurencePlan/PlatinumPlan.java
@@ -1,0 +1,15 @@
+package kdu.ques4.InsurencePlan;
+
+import kdu.ques4.HealthInsurancePlan;
+
+public class PlatinumPlan extends HealthInsurancePlan {
+    public PlatinumPlan(){
+        setCoverage(0.9);
+        setDiscount(50.0);
+    }
+
+    @Override
+    public double computeMonthlyPremium(double salary, int age, boolean smoking) {
+        return 0.08 * salary + getOfferedBy().computeMonthlyPremium(this, age, smoking);
+    }
+}
