+++ backend/homework/homework-3/src/main/java/kdu/ques4/Main.java
@@ -1,0 +1,16 @@
+package kdu.ques4;
+
+import kdu.ques4.InsurencePlan.PlatinumPlan;
+
+public class Main {
+    public static void main(String[] args){
+        User staff = new User();
+        InsuranceBrand insuranceBrand = new BlueCrossBlueShield();
+        HealthInsurancePlan insurancePlan = new PlatinumPlan();
+
+        insurancePlan.setOfferedBy(insuranceBrand);
+        staff.setInsurancePlan(insurancePlan);
+
+        System.out.println(insurancePlan.computeMonthlyPremium(5000, 56, true));
+    }
+}
